requirements.txt
#llangchain
#llangchain-community
#langchain-google-vertexai
#langchain-openai
#lpillow
#psycopg[binary]
psycopg2-binary
asyncpg
greenlet
sqlalchemy
passlib
bcrypt
python-multipart
python-jose
python-dotenv
fastapi
uvicorn